Introduce simple const vars evaluation logic
