+++ src/types/index.ts
@@ -1,8 +1,9 @@
 export * from "./ast";
 export * from "./builtins";
-export * from "./typeStrings";
-export * from "./utils";
+export * from "./eval_const";
 export * from "./infer";
+export * from "./misc";
 export * from "./polymorphic";
 export * from "./reserved";
-export * from "./eval_const";
+export * from "./typeStrings";
+export * from "./utils";
