+++ src/types/misc.ts
@@ -1,0 +1,1 @@
+export class SolTypeError extends Error {}
